requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.51
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==2.3.8
waitress==2.1.2
python-dotenv==1.0.0
Flask-Limiter==3.5.0
limits==3.6.0
cryptography==41.0.4
bcrypt==4.0.1
WTForms==3.1.1
mysqlclient==2.2.0
alembic==1.12.1
